requirements.txt
playsound
openai
baidu-aip
requests
pyaudio
jieba
torch
soundfile
bs4
numpy
scipy
unidecode
phonemizer
pypinyin
Cython
selenium
pycryptodome